validation endpoint fix
