+++ Presentation.Web/Presentation.Web.csproj
@@ -380,6 +380,7 @@
     <TypeScriptCompile Include="app\models\ui-customization\config\blueprints\it-contract-ui-customization-config-blueprint.ts" />
     <TypeScriptCompile Include="app\helpers\KendoToolbarCustomizationHelper.ts" />
     <TypeScriptCompile Include="app\helpers\UiCustomizationHelper.ts" />
+    <Compile Include="Controllers\API\V1\ItSystemUsageValidationController.cs" />
     <Compile Include="Controllers\API\V1\KendoOrganizationalConfigurationController.cs" />
     <Compile Include="Controllers\API\V1\Mapping\DTOMappingExtensions.cs" />
     <Compile Include="Controllers\API\V1\OData\ItContractOverviewReadModelsController.cs" />
